Добавлена проверка введенных пользователем данных для фиксирования количества шагов.

+++ src/Converter.java
@@ -5,14 +5,21 @@ public class Converter {
 
     /**
      * Конвертирует шаги в километры.
-      * @param steps количество шагов.
+     *
+     * @param steps количество шагов.
      * @return количество километров.
      */
-    static double convertDistance(int steps) {
-        return (double) (steps * 75 / 100000);
+    static int convertDistance(int steps) {
+        return steps * 75 / 100000;
     }
 
-    static double convertCalories(int steps) {
-        return (double) (steps * 50 / 1000);
+    /**
+     * Конвертирует шаги в сожженные килокалории.
+     *
+     * @param steps количество шагов.
+     * @return количество сожженных килокалорий.
+     */
+    static int convertCalories(int steps) {
+        return steps * 50 / 1000;
     }
 }
